Fix/upload bug

+++ covidml/routes.py
@@ -47,8 +47,13 @@ def upload_page():
             results['activity'] = activity
             results['confidence'] = confidence
             results['ad'] = ad_analysis
+            
+            if not os.path.exists('covidml/static/temp/'):
+                os.mkdir('covidml/static/temp/')
+            
+            if os.path.exists('covidml/static/temp/results.csv'):
+                os.remove('covidml/static/temp/results.csv')
                 
-            os.remove('covidml/static/temp/results.csv')
             df = pd.DataFrame(results, index=[0])
             df.to_csv('covidml/static/temp/results.csv', index=False)
                 
@@ -58,6 +63,7 @@ def upload_page():
             results['adImage'] = ad_img
             
     except Exception as e:
+        print(e)
         flash("Invalid SMILES, enter a valid SMILES to make a prediction", category="danger")
                 
     return render_template('upload.html', title='Upload', form=form, file_form=UploadFile(), results=results)
@@ -88,8 +94,6 @@ def upload_file_page():
                 if l.strip():
                     lines += 1
         
-        all_results = []
-        df = pd.DataFrame()
         for line in range(lines):
             with open(f.filename, 'r') as file:
                 content = file.readlines()
@@ -122,13 +126,18 @@ def upload_file_page():
                 results['confidence'] = confidence
                 results['ad'] = ad_analysis
 
-                df_temp = pd.DataFrame(all_results)
-                pd.concat([df, df_temp], axis=0)
+                if not os.path.exists('covidml/static/temp/'):
+                    os.mkdir('covidml/static/temp/')
+            
+                if os.path.exists('covidml/static/temp/results.csv'):
+                    os.remove('covidml/static/temp/results.csv')
+                
+                df = pd.DataFrame(results, index=[0])
+                df.to_csv('covidml/static/temp/results.csv', index=False)              
+                                
                 results['model_type'] = model_type
                 results['image'] = structure_img
                 results['adImage'] = ad_img
-                
-                all_results.append(results)
             
             except Exception as e:
                 flash("Invalid SMILES provided in text file", "danger")
@@ -136,11 +145,8 @@ def upload_file_page():
             
         # remove uploaded file
         os.remove(f.filename)
-            
-        os.remove('covidml/static/temp/results.csv')
-        df.to_csv('covidml/static/temp/results.csv', index=False)
-        print(len(all_results))
-        return render_template('upload_file.html', title='Results', file_form=file_form, all_results=all_results)
+
+        return render_template('upload_file.html', title='Results', file_form=file_form, results=results)
     return render_template('upload_file.html', title='Upload File', file_form=file_form)
